Move HandoffProtocol to shared module
